Refactor: 'Samarbeid' funksjoner -> SamarbeidTransactional
